% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{bc_date}
\alias{bc_date}
\title{Package: \code{eqplot}}
\source{
Dataset NOAA Significant Earthquake Database
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
National Geophysical Data Center / World Data Service (NGDC / WDS):
Significant Earthquake Database.
National Geophysical Data Center, NOAA. doi:10.7289 / V5TD9V7K
Get BC date from mirrored AD date: \code{bc_date()}

This sub function converts a mirrored AD date to BC date and internally used.
}
\usage{
bc_date(ad)
}
\arguments{
\item{ad}{A Date class of mirrored AD date , a positive Date type (e.g. "2001-01-01").}
}
\value{
A Date class of BC date, a negative Date type (e.g. "-2001-01-01")
 corresponding to the mirrored AD date
}
\description{
This package will provide tools to visualize the earthquake dataset obtained
from the U.S. National Oceanographic and Atmospheric Administration (NOAA) on
significant earthquakes around the world. This dataset contains information
about 5,933 earthquakes over an approximately 4,000 year time span.
}
\note{
As this is just an internal function to be used for the other functions,
 it is not exported.
}
