% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean data function: \code{eq_clean_data()}}
\usage{
eq_clean_data(raw_df)
}
\arguments{
\item{raw_df}{A data frame of raw data.}
}
\value{
A data frame of clean data
}
\description{
This function takes raw NOAA data frame and returns a clean data frame.
The clean data frame should have the following:
  A date column created by uniting the year, month, day and converting it
   to the Date class,
   LATITUDE, LONGITUDE, EQ_PRIMARY(magnitude) and TOTAL_DEATHS columns
   converted to numeric class.
   The rows whose LATITUDE, LONGITUDE or EQ_PRIMARY is missing (NA) are
   removed.
}
\examples{
\dontrun{
data(raw_df)
clean_df <- eq_clean_data(raw_df)
}

}
