% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create annotation label: \code{eq_create_label()}}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data frame of the clean data}
}
\value{
An HTML label annotating each point with the cleaned location,
 the magnitude (EQ_PRIMARY), and the total number of deaths (\code{TOTAL_DEATHS})
 in pop up window.
}
\description{
This function takes the dataset as an argument and creates an HTML label
 that can be used as the annotation text in the leaflet map. This function
 should put together a character string for each earthquake that will show
 the cleaned location (as cleaned by the \code{eq_location_clean()} function
 created above, the magnitude (EQ_PRIMARY), and the total number of deaths
 (\code{TOTAL_DEATHS}), with boldface labels for each ("Location", "Total
 deaths", and "Magnitude"). If an earthquake is missing values for any of
 these, both the label and the value should be skipped for that element of
 the tag.
}
\examples{
\dontrun{
library(dplyr)
data(raw_data)
raw_data \%>\% eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}

}
