% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_date}
\alias{eq_date}
\title{Get Date class date from chracter type date reprezentation: \code{eq_date}}
\usage{
eq_date(x)
}
\arguments{
\item{x}{A chracter string reprezenting date in the format of "yyyy-mm-dd"
or "-yyyy-mm-dd".}
}
\value{
A Date class date corresponding to "yyyy-mm-dd" or "-yyyy-mm-dd".
}
\description{
This function converts chracter type date to Date type date.
}
\note{
This can be used to specify the date of earthquake occurence
for the other functions` arguments.
}
\examples{
eq_date("2001-01-01")
eq_date("-2001-01-01")


}
