% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean the LOCATION_NAME : \code{eq_location_clean()}}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{A chracter vector}
}
\value{
A character vector cleaned
}
\description{
This function cleans the LOCATION_NAME column by stripping out the country
 name (including the colon) and converts names to title case (as opposed to
 all caps).
}
\examples{
\dontrun{
 library(readr)
 # Note that the raw data set has to be downloaded beforehand
 raw_df <- readr::read_delim("data/rawdata.txt", "\\t")
 raw_df <- raw_df \%>\% dplyr::mutate(LOCATION_NAME = eq_location_clean(LOCATION_NAME))
}

}
