% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_map}
\alias{eq_map}
\title{Visualize earthquakes in space: \code{eq_map()}}
\usage{
eq_map(data, annot_col = "DATE", cluster = FALSE)
}
\arguments{
\item{data}{A data frame of the clean data}

\item{annot_col}{A character string of a column name of the \code{data} for
annotation}

\item{cluster}{A logical flag (deafult FALSE). If TRUE, ClusterOPtion is used.}
}
\value{
Visualization of earthquakes in space annotating each point with
in pop up window
}
\description{
This function takes an argument data containing the filtered data frame
with earthquakes to visualize. The function maps the epicenters (LATITUDE
/LONGITUDE) and annotates each point with in pop up window containing
annotation data stored in a column of the data frame. The user should be
able to choose which column is used for the annotation in the pop-up with
a function argument named \code{annot_col}. Each earthquake should be shown with
a circle, and the radius of the circle should be proportional to the
earthquake's magnitude (\code{EQ_PRIMARY}).
}
\examples{
\dontrun{
library(dplyr)
data(raw_df)
raw_df \%>\% eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")
}

}
