% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\docType{data}
\name{eq_theme}
\alias{eq_theme}
\title{New theme for the timeline plot: \code{eq_theme}}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 60.}
\usage{
eq_theme
}
\description{
This theme is for the earthquake timeline plot
}
\keyword{datasets}
