% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{geom_timeline}
\alias{geom_timeline}
\title{GeomTimeline: \code{geom_timeline()}}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "timeline",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, xmin = NULL, xmax = NULL, ...)
}
\arguments{
\item{data}{A data frame of cleaned data}

\item{xmin}{A Date class date which specifies the minimum timeline range.
This is optional. If not specified, The minimum timeline range of the
\code{data} is assumed.}

\item{xmax}{A Date class date which specifies the maximum timeline range.
This is optional. If not specified, The maximum timeline range of the
 \code{data} is assumed.}

\item{aes(x, }{y, size, color, fill) Aesthetics as follows:
x  A Date class (DATE)
y  A factor (e.g. COUNTRY)
size   A numeric magnitude (EQ_PRIMARY)
color  A numeric deaths (TOTAL_DEATHS)
fill   A numeric deaths (TOTAL_DEATHS)}

\item{alpha}{A numeric value to specify transparency of points.
THis is optional (default value is 0.3)}
}
\value{
plot of a time line of earthquakes ranging from xmin to xmax dates
         with a point for each earthquake.
}
\description{
This geom and geom_timeline function will plot a time line of earthquakes
ranging from xmin to xmax dates with a point for each earthquake.
Optional aesthetics include color, size, and alpha (for transparency).
The x aesthetic is a date and an optional y aesthetic is a factor indicating
some stratification in which case multiple time lines will be plotted for
each level of the factor (e.g. country).
}
\examples{
\dontrun{
data(raw_df)
test_df <- raw_df \%>\% eq_clean_data() \%>\%
filter(COUNTRY == "TURKEY" | COUNTRY == "JAPAN") \%>\%
dplyr::filter(DATE >= eq_date("2000-01-01") & DATE <= eq_date("2017-01-01"))
ggtimeline <- ggplot2::ggplot(test_df)
ggtimeline2 <- ggtimeline +
               geom_timeline(ggplot2::aes(
                           x = DATE,
                           y = COUNTRY,
                           size = EQ_PRIMARY,
                           color = TOTAL_DEATHS,
                           fill = TOTAL_DEATHS) +
             eq_theme +
             ggplot2::scale_size_continuous(name = "Richer scale value", breaks = c(0, 2, 4, 6, 8)) +
             ggplot2::scale_color_continuous(name = "# deaths") +
             ggplot2::scale_fill_continuous(name = "# deaths") +
             ggplot2::labs(title = "Earthquakes")
ggtimeline2
}

}
