% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{GeomTimeline_label:  \code{geom_timeline_label()}}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "timelinelabel",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, xmin = NULL, xmax = NULL, n_max = 3, ...)
}
\arguments{
\item{data}{A data frame of cleaned data}

\item{xmin}{A Date class date which specifies the minimum timeline range.
This is optional. If not specified, the minimum timeline range of the
\code{data} is assumed.}

\item{xmax}{A Date class date which specifies the maximum timeline range.
This is optional. If not specified, the maximum timeline range of the
 \code{data} is assumed.}

\item{n_max}{A integer to subset to n_max number of earthquakes.
This is optional. If not specified, 3 is assumed (default).}

\item{aes(x, }{y, mag, label) Aesthetics as follows:
x     A Date class column (DATE)
y     A factor column (e.g. COUNTRY)
mag   A numeric column (e.g. EQ_PRIMARY)
label A character column (e.g LOCATION_NAME)}
}
\value{
Adds a vertical line to each data point with a text annotation
         for \code{n_max} largest (by magnitude) earthquakes.
}
\description{
This geom and geom_timeline_label function will add annotations to the
earthquake data. This geom adds a vertical line to each data point with
a text annotation (e.g. the location of the earthquake) attached to each
line. There should be an option to subset to n_max number of earthquakes,
where we take the n_max largest (by magnitude) earthquakes. Aesthetics are
x, which is the date of the earthquake and label which takes the column name
from which annotations will be obtained.
}
\note{
This geom_timeline_label function should be added with (+) to the previous
      geom_timeline()  or ggplot2 object generated by geom_timeline() with the
      same timeline range as the previous geom_timeline(). Unfortunately at this
      version it is required for users to ensure the coincident of the timeline
      range.
}
\examples{
\dontrun{
ggtimeline2 + geom_timeline_label(ggplot2::aes(
                                  x = DATE,
                                  y = COUNTRY,
                                  mag = EQ_PRIMARY,
                                  label = LOCATION_NAME),
                                  n_max = 5)
}

}
