% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_df}
\alias{raw_df}
\title{Significant earthquakes around the world}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5979 rows and 47 columns.}
\usage{
raw_df
}
\description{
This dataset contains information about 5,933 earthquakes over
  an approximately 4,000 year time span.
}
\details{
@source Dataset NOAA Significant Earthquake Database
  \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
  National Geophysical Data Center / World Data Service (NGDC / WDS):
  Significant Earthquake Database.
  National Geophysical Data Center, NOAA. doi:10.7289 / V5TD9V7K

@format A data frame with columns:
  \describe{
  ID
  Flag Tsunami
  Date: Year, Month, Day, Hour, Minute, Second
  Earthquake Location:
              Country
              State,
              Location Name
              Latitude
              Longitude
              Region Code
  Focal Depth
  Magnitude: MS, MW, mb, ML, MFA, Unknown
  Modified Mercalli Intensity (MMI)
  Earthquake Effects:
              Number of Deaths, Deaths Desccription
              Number of Missing, Missing Description
              Number of Injuries, Injuries Description
              Damage Millions of Dollars, Damage Description
              Number of Houses Destroyed, Houses Destroyed Description
              Number of Houses Damaged, Houses Damaged Description
  Total Effects (Earthquake and Tsunami, Volcano, etc.):
              Total Number of Deaths
              Total Deaths Desccription
              Total Number of Missing
              Total Missing Description
              Total Number of Injuries
              Total Injuries Description
              Total Damage in Millions of Dollars
              Total Damage Description
              Total Number of Houses Destroyed
              Total Houses Destroyed Description
              Total Number of Houses Damaged
              Total Houses Damaged Description
 }
}
\keyword{datasets}
